Initial common code port to 1.21.5

+++ src/main/java/cf/witcheskitchen/api/entity/WKBroomBase.java
@@ -61,7 +61,7 @@ public abstract class WKBroomBase extends Entity {
     }
 
     @Override
-    public boolean handleFallDamage(float fallDistance, float damageMultiplier, DamageSource damageSource) {
+    public boolean handleFallDamage(double fallDistance, float damageMultiplier, DamageSource damageSource) {
         return false;
     }
 }
